requirements.txt
streamlit>=1.35.0
openai>=1.51.0
requests>=2.31.0
google-cloud-logging
google-cloud-firestore
google-cloud-storage
Pillow
markdown-pdf
trimesh
scikit-image
beautifulsoup4>=4.12.2
Jinja2>=3.1.0
pytest-timeout>=2.3.1
pymupdf>=1.24.0
numpy>=1.26
faiss-cpu>=1.7.4; platform_system != "Windows"
PyYAML>=6.0
filelock>=3.0
